Clippy 1.49.0 updates and dht persistence test fix (#2156)
## Issue Addressed

`test_dht_persistence` failing

## Proposed Changes

Bind `NetworkService::start` to an underscore prefixed variable rather than `_`.  `_` was causing it to be dropped immediately

This was failing 5/100 times before this update, but I haven't been able to get it to fail after updating it

+++ testing/remote_signer_test/src/utils.rs
@@ -102,11 +102,8 @@ pub fn get_block<E: EthSpec>(seed: u64) -> BeaconBlock<E> {
         signature: Signature::empty(),
     };
     let indexed_attestation: IndexedAttestation<E> = IndexedAttestation {
-        attesting_indices: VariableList::new(vec![
-            0 as u64;
-            E::MaxValidatorsPerCommittee::to_usize()
-        ])
-        .unwrap(),
+        attesting_indices: VariableList::new(vec![0_u64; E::MaxValidatorsPerCommittee::to_usize()])
+            .unwrap(),
         data: AttestationData::default(),
         signature: AggregateSignature::empty(),
     };
